Split game logic into separate module
