+++ gomoku/core/__init__.py
@@ -1,1 +1,6 @@
 """ゲームの基本機能を提供するサブパッケージ"""
+
+from .game import Gomoku, count_chains_open_ends
+from .gomoku_env import GomokuEnv
+
+__all__ = ["Gomoku", "GomokuEnv", "count_chains_open_ends"]
