+++ gomoku/core/gomoku_env.py
@@ -1,157 +1,8 @@
 import numpy as np
 
 # 同一パッケージ内のユーティリティを相対インポート
+from .game import Gomoku, count_chains_open_ends
 from .utils import opponent_player
-
-
-def _scan_chain_open_end(
-    board: np.ndarray,
-    x: int,
-    y: int,
-    dx: int,
-    dy: int,
-    player: int,
-):
-    """指定座標から1方向に伸びる連の長さと両端の空き状況を返すヘルパー"""
-
-    board_size = board.shape[0]
-
-    # 逆方向に同じ色の石がある場合、ここは連の始点ではない
-    prev_x, prev_y = x - dx, y - dy
-    if 0 <= prev_x < board_size and 0 <= prev_y < board_size:
-        if board[prev_x, prev_y] == player:
-            return 0, 0
-
-    # (x, y) から (dx, dy) 方向に連を伸ばす
-    length = 1
-    cx, cy = x + dx, y + dy
-    while 0 <= cx < board_size and 0 <= cy < board_size and board[cx, cy] == player:
-        length += 1
-        cx += dx
-        cy += dy
-
-    # 両端が空いているかを確認
-    open_ends = 0
-    if 0 <= prev_x < board_size and 0 <= prev_y < board_size:
-        if board[prev_x, prev_y] == 0:
-            open_ends += 1
-    if 0 <= cx < board_size and 0 <= cy < board_size:
-        if board[cx, cy] == 0:
-            open_ends += 1
-
-    return length, open_ends
-
-def count_chains_open_ends(board: np.ndarray, player: int):
-    """
-    board上で指定playerの
-    - ちょうど2連 / 3連 / 4連
-      それぞれについて
-       (open2: 両端空, open1: 片端のみ空) の本数を数える。
-
-    返り値は辞書形式で:
-    {
-      2: {'open2': 個数, 'open1': 個数},
-      3: {'open2': 個数, 'open1': 個数},
-      4: {'open2': 個数, 'open1': 個数}
-    }
-    のようにまとめて返す。
-    """
-    board_size = board.shape[0]
-    directions = [(1, 0), (0, 1), (1, 1), (1, -1)]
-
-    result = {
-        2: {'open2': 0, 'open1': 0},
-        3: {'open2': 0, 'open1': 0},
-        4: {'open2': 0, 'open1': 0},
-    }
-
-    for x in range(board_size):
-        for y in range(board_size):
-            if board[x, y] != player:
-                continue
-
-            for dx, dy in directions:
-                length, open_ends = _scan_chain_open_end(board, x, y, dx, dy, player)
-
-                if length in [2, 3, 4]:
-                    if open_ends == 2:
-                        result[length]['open2'] += 1
-                    elif open_ends == 1:
-                        result[length]['open1'] += 1
-
-    return result
-
-
-class Gomoku:
-    """
-    五目並べのゲームロジック。
-    board_size x board_size の盤面を扱い、
-      - 盤面の初期化
-      - 石を置けるかどうかのチェック
-      - 石を置いたあとの勝敗判定
-    などを行う。
-    """
-    def __init__(self, board_size=9):
-        self.board_size = board_size
-        self.board = np.zeros((board_size, board_size), dtype=np.int8)
-        self.current_player = 1  # 先手=1(黒), 後手=2(白)
-
-    def reset(self):
-        """盤面をリセットして先手を黒番(1)にする。"""
-        self.board = np.zeros((self.board_size, self.board_size), dtype=np.int8)
-        self.current_player = 1
-
-    def is_valid_move(self, x, y):
-        """(x, y)が盤内であり、かつ空マスならTrue"""
-        if x < 0 or x >= self.board_size or y < 0 or y >= self.board_size:
-            return False
-        return (self.board[x, y] == 0)
-
-    def place_stone(self, x, y):
-        """
-        (x, y)に現在のプレイヤーの石を置き、
-        手番を相手プレイヤーに交替。
-        """
-        self.board[x, y] = self.current_player
-        # 石を置いたら次の手番へ交代
-        self.current_player = opponent_player(self.current_player)
-
-    def check_winner(self):
-        """
-        盤面を見て、勝者がいるかどうかを判定する。
-        - 1(黒)勝ちなら 1
-        - 2(白)勝ちなら 2
-        - 引き分けなら -1
-        - まだ決着ついていなければ 0
-        """
-        directions = [(1,0), (0,1), (1,1), (1,-1)]
-        for x in range(self.board_size):
-            for y in range(self.board_size):
-                player = self.board[x, y]
-                if player == 0:
-                    continue
-                for dx, dy in directions:
-                    if self._count_stones(x, y, dx, dy, player) >= 5:
-                        return player
-        # 盤が全て埋まっていたら引き分け
-        if np.all(self.board != 0):
-            return -1
-        return 0
-
-    def _count_stones(self, x, y, dx, dy, player):
-        """
-        (x,y)から(dx,dy)方向に連続している同色(player)の石を数える。
-        """
-        count = 0
-        cx, cy = x, y
-        while 0 <= cx < self.board_size and 0 <= cy < self.board_size:
-            if self.board[cx, cy] == player:
-                count += 1
-                cx += dx
-                cy += dy
-            else:
-                break
-        return count
 
 
 class GomokuEnv:
